chore: hoist `spawnPromisified` to repo root
