+++ recipes/correct-ts-specifiers/src/logger.test.ts
@@ -2,7 +2,7 @@ import assert from 'node:assert/strict';
 import { execPath } from 'node:process';
 import { describe, it } from 'node:test';
 
-import { spawnPromisified } from '../test/spawn-promisified.ts';
+import { spawnPromisified } from '../../../build/spawn-promisified.ts';
 
 describe('logger', () => {
 	it('should emit non-error entries to standard out, collated by source module', async (t) => {
